feat: lock down deleteItem permissions

+++ backend/src/resolvers/Mutation.js
@@ -43,7 +43,16 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
-    const item = await ctx.db.query.item({ where }, `{ id title}`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`)
+    // Check if the user own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    )
+
+    if (!hasPermissions || !ownsItem) {
+      throw new Error("You don't have permission to do that!")
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info)
   },
